cpytraceafl_record_loc: change recording scheme to perform right-shift before storing prev_loc
this is in line with how AFL's instrumentation generally uses prev_loc, so
we shouldn't end up double shifting or no-shifting when jumping between a
natively-instrumented context and a cpytraceafl context. this should also
make it easier to present a straightforward explanation of how the
function uses this_loc.

i was initially surprised that this didn't break any ngram-using test
results but it makes sense if you think about it.